chore: Merge DosPathResolver and DosDriveManager into one file
